Updated Jenkinsfile for automatic deployment and notifications
Added environment variables for service, image, and repository names, and updated the pipeline logic to automate deployments based on branch name. The changes will control the version creation, docker build, and push sequence based on whether the change is in the 'master' or a feature branch. This will enable easy maintenance and more control over the deployment process.

Also added steps to send notifications at the start and end of each build and after successful service updates. This will help to monitor the builds and deployments more efficiently. Some minor tweaks were also made to formatting of version number strings.